Added PropTypes for EventDetails and SendInviteModal
