+++ src/components/event-calendar/EventCalendar.js
@@ -3,7 +3,9 @@ import moment from 'moment'
 import { Panel, Grid, Row, Col } from 'react-bootstrap'
 import BigCalendar from 'react-big-calendar'
 import 'react-big-calendar/lib/css/react-big-calendar.css'
+
 import { uuid } from '../../flow/io/eventFN'
+
 let localizer = BigCalendar.momentLocalizer(moment)
 let allViews = Object.keys(BigCalendar.Views).map(k => BigCalendar.Views[k])
 
@@ -120,6 +122,7 @@ class EventCalendar extends Component {
       shareUrl =
         window.location.origin + '/?intent=addics&url=' + publicCalendar
     }
+
     const calendarView = (
       <BigCalendar
         localizer={localizer}
